+++ Model/IMConnectors.m
@@ -77,8 +77,8 @@
   else
     {
       const char	*type;
-      unsigned int	size;
-      int		offset;
+      size_t		size;
+      ptrdiff_t		offset;
 
       /*
        * Use the GNUstep additional function to set the instance
@@ -101,8 +101,8 @@
   else
     {
       const char	*type;
-      unsigned int	size;
-      int		offset;
+      size_t		size;
+      ptrdiff_t		offset;
 
       /*
        * Use the GNUstep additional function to set the instance
@@ -159,8 +159,8 @@
     {
       const char	*nam = [label cString];
       const char	*type;
-      unsigned int	size;
-      int		offset;
+      size_t		size;
+      ptrdiff_t		offset;
 
       /*
        * Use the GNUstep additional function to set the instance
